Minor update, added submodules xstring and utopia, updated build scripts.

+++ build.sh
@@ -3,29 +3,48 @@
 exe=xc
 src=src/*.c
 cc=gcc
-std=-std=c99
-opt=-O2
 inc=-I.
+ldir=-Llib
 
-wflags=(
+libs=(
+    -lutopia 
+    -lxstring
+)
+
+flags=(
+    -std=c99
     -Wall 
     -Wextra
+    -O2
 )
 
-cflags=(
-    ${std[*]} 
-    ${opt[*]} 
-    ${wflags[*]}
-    ${inc[*]}
-)
+buildlib() {
+    pushd $1 && ./build.sh $2 && mv *.a ../lib/ && popd
+}
+
+build() {
+    [ ! -d lib ] && mkdir lib
+    buildlib utopia -slib
+    buildlib xstring static
+}
 
 comp() {
-    $cc ${cflags[*]} $src -o $exe
+    $cc ${flags[*]} $inc $ldir ${libs[*]} $src -o $exe
+}
+
+clean() {
+    [ -d lib ] && rm -r lib && echo "Deleted 'lib'."
+    [ -f $exe ] && rm $exe && echo "Deleted '$exe'."
+    return 0
 }
 
 case "$1" in
-    "clean")
-        rm $exe;;
-    *)
+    "build")
+        build;;
+    "comp")
         comp;;
+    "clean")
+        clean;;
+    *)
+        echo "Use with 'build', 'comp' or 'clean'";;
 esac
